requirements.txt
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.2.0
blinker==1.8.2
cachelib==0.13.0
certifi==2020.6.20
click==8.1.7
command-not-found==0.3
contourpy==1.3.1
cryptography==3.4.8
cycler==0.12.1
dbus-python==1.2.18
Deprecated==1.2.14
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.7.0
EasyProcess==1.1
email_validator==2.2.0
entrypoint2==1.1
exceptiongroup==1.3.0
filelock==3.6.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
flask-marshmallow==1.2.1
Flask-RESTful==0.3.10
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.56.0
greenlet==3.1.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.20.2
httpx==0.28.1
idna==3.10
img2pdf==0.5.1
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
jiter==0.11.1
joblib==1.5.2
keyring==23.5.0
kiwisolver==1.4.8
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==5.3.0
MarkupSafe==2.1.5
marshmallow==3.22.0
marshmallow-sqlalchemy==1.1.0
matplotlib==3.10.0
more-itertools==8.10.0
MouseInfo==0.1.3
msgspec==0.18.6
mss==10.0.0
netifaces==0.11.0
numpy==2.2.2
oauthlib==3.2.0
openai==2.7.1
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.1
pandas==2.3.3
pikepdf==9.2.1
pillow==11.1.0
pipenv==11.9.0
platformdirs==2.5.1
PyAutoGUI==0.9.54
pydantic==2.12.3
pydantic_core==2.41.4
PyGetWindow==0.0.9
PyGObject==3.42.1
PyJWT==2.3.0
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.9.0
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==1.0.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-xlib==0.15
pytweening==1.2.0
pytz==2024.2
PyYAML==5.4.1
scikit-learn==1.7.2
scipy==1.15.3
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.34
systemd-python==234
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
virtualenv==20.13.0+ds
virtualenv-clone==0.3.0
wadllib==1.3.6
Werkzeug==3.0.4
wrapt==1.16.0
WTForms==3.1.2
zipp==1.0.0
